+++ packages/xchain-thorchain/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v.x.x.x 
+
+### Update
+
+- update `getTransactionHistory`, `transfer`, `deposit` to support synth assets.
+
 # v.0.13.1 (2021-03-18)
 
 ### Fix